fix(provider): subscribe to stream from provider
with a stream subscribed to from within a consumer, every time a wrapped component is unmounted and
mounted it is initialised with the initial state from the store, not the most recent state. To fix
this, we wrap the Provider created from createContext with a component which creates the
subscription, and then passes down the entire state to the Provider. This way any Consumers get the
latest state from the Provider, instead of the initial state from the store

BREAKING CHANGE: Provider now requires a 'store' property instead of 'value'